pyproject.toml
[project]
name = "taskapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[all]>=0.115.13",
    "middleware>=1.2.3",
    "passlib[argon2]>=1.7.4",
    "pydantic[email]>=2.11.7",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
